save function restricted to completion of required fields
